+++ opentelemetry-sdk/src/metrics/exporter.rs
@@ -1,11 +1,93 @@
 //! Interfaces for exporting metrics
 
-use crate::error::OTelSdkResult;
-use std::time::Duration;
+use opentelemetry::InstrumentationScope;
 
-use crate::metrics::data::ResourceMetrics;
+use crate::{error::OTelSdkResult, Resource};
+use std::{fmt::Debug, slice::Iter, time::Duration};
 
-use super::Temporality;
+use super::{
+    data::Metric,
+    reader::{ResourceMetricsData, ScopeMetricsData},
+    Temporality,
+};
+
+/// A collection of [`BatchScopeMetrics`] and the associated [Resource] that created them.
+#[derive(Debug)]
+pub struct ResourceMetrics<'a> {
+    /// The entity that collected the metrics.
+    pub resource: &'a Resource,
+    /// The collection of metrics with unique [InstrumentationScope]s.
+    pub scope_metrics: ScopeMetricsLendingIter<'a>,
+}
+
+/// Iterator over libraries instrumentation scopes ([`InstrumentationScope`]) together with metrics.
+/// Doesn't implement standard [`Iterator`], because it returns borrowed items. AKA "LendingIterator".
+pub struct ScopeMetricsLendingIter<'a> {
+    iter: Iter<'a, ScopeMetricsData>,
+}
+
+/// A collection of metrics produced by a [`InstrumentationScope`] meter.
+#[derive(Debug)]
+pub struct ScopeMetrics<'a> {
+    /// The [InstrumentationScope] that the meter was created with.
+    pub scope: &'a InstrumentationScope,
+    /// The list of aggregations created by the meter.
+    pub metrics: MetricsLendingIter<'a>,
+}
+
+/// Iterator over aggregations created by the meter.
+/// Doesn't implement standard [`Iterator`], because it returns borrowed items. AKA "LendingIterator".
+pub struct MetricsLendingIter<'a> {
+    iter: Iter<'a, Metric>,
+}
+
+impl<'a> ResourceMetrics<'a> {
+    pub(crate) fn new(rm: &'a ResourceMetricsData) -> Self {
+        Self {
+            resource: &rm.resource,
+            scope_metrics: ScopeMetricsLendingIter {
+                iter: rm.scope_metrics.iter(),
+            },
+        }
+    }
+}
+
+impl<'a> ScopeMetrics<'a> {
+    fn new(sm: &'a ScopeMetricsData) -> Self {
+        Self {
+            scope: &sm.scope,
+            metrics: MetricsLendingIter {
+                iter: sm.metrics.iter(),
+            },
+        }
+    }
+}
+
+impl Debug for ScopeMetricsLendingIter<'_> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("BatchScopeMetrics").finish()
+    }
+}
+
+impl ScopeMetricsLendingIter<'_> {
+    /// Advances the iterator and returns the next value.
+    pub fn next(&mut self) -> Option<ScopeMetrics<'_>> {
+        self.iter.next().map(ScopeMetrics::new)
+    }
+}
+
+impl MetricsLendingIter<'_> {
+    /// Advances the iterator and returns the next value.
+    pub fn next(&mut self) -> Option<&Metric> {
+        self.iter.next()
+    }
+}
+
+impl Debug for MetricsLendingIter<'_> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("BatchMetrics").finish()
+    }
+}
 
 /// Exporter handles the delivery of metric data to external receivers.
 ///
@@ -18,7 +100,7 @@ pub trait PushMetricExporter: Send + Sync + 'static {
     /// considered unrecoverable and will be logged.
     fn export(
         &self,
-        metrics: &mut ResourceMetrics,
+        metrics: ResourceMetrics<'_>,
     ) -> impl std::future::Future<Output = OTelSdkResult> + Send;
 
     /// Flushes any metric data held by an exporter.