+++ opentelemetry-sdk/src/metrics/in_memory_exporter.rs
@@ -1,7 +1,5 @@
 use crate::error::{OTelSdkError, OTelSdkResult};
-use crate::metrics::data::{
-    ExponentialHistogram, Gauge, Histogram, MetricData, ResourceMetrics, Sum,
-};
+use crate::metrics::data::{ExponentialHistogram, Gauge, Histogram, MetricData, Sum};
 use crate::metrics::exporter::PushMetricExporter;
 use crate::metrics::Temporality;
 use crate::InMemoryExporterError;
@@ -10,7 +8,9 @@ use std::fmt;
 use std::sync::{Arc, Mutex};
 use std::time::Duration;
 
-use super::data::{AggregatedMetrics, Metric, ScopeMetrics};
+use super::data::{AggregatedMetrics, Metric};
+use super::exporter::ResourceMetrics;
+use super::reader::{ResourceMetricsData, ScopeMetricsData};
 
 /// An in-memory metrics exporter that stores metrics data in memory.
 ///
@@ -60,7 +60,7 @@ use super::data::{AggregatedMetrics, Metric, ScopeMetrics};
 ///# }
 /// ```
 pub struct InMemoryMetricExporter {
-    metrics: Arc<Mutex<VecDeque<ResourceMetrics>>>,
+    metrics: Arc<Mutex<VecDeque<ResourceMetricsData>>>,
     temporality: Temporality,
 }
 
@@ -146,11 +146,35 @@ impl InMemoryMetricExporter {
     /// let exporter = InMemoryMetricExporter::default();
     /// let finished_metrics = exporter.get_finished_metrics().unwrap();
     /// ```
-    pub fn get_finished_metrics(&self) -> Result<Vec<ResourceMetrics>, InMemoryExporterError> {
+    pub fn get_finished_metrics(&self) -> Result<Vec<ResourceMetricsData>, InMemoryExporterError> {
         let metrics = self
             .metrics
             .lock()
-            .map(|metrics_guard| metrics_guard.iter().map(Self::clone_metrics).collect())
+            .map(|metrics_guard| {
+                metrics_guard
+                    .iter()
+                    .map(|data| ResourceMetricsData {
+                        resource: data.resource.clone(),
+                        scope_metrics: data
+                            .scope_metrics
+                            .iter()
+                            .map(|data| ScopeMetricsData {
+                                scope: data.scope.clone(),
+                                metrics: data
+                                    .metrics
+                                    .iter()
+                                    .map(|data| Metric {
+                                        name: data.name.clone(),
+                                        description: data.description.clone(),
+                                        unit: data.unit.clone(),
+                                        data: Self::clone_data(&data.data),
+                                    })
+                                    .collect(),
+                            })
+                            .collect(),
+                    })
+                    .collect()
+            })
             .map_err(InMemoryExporterError::from)?;
         Ok(metrics)
     }
@@ -172,26 +196,26 @@ impl InMemoryMetricExporter {
             .map(|mut metrics_guard| metrics_guard.clear());
     }
 
-    fn clone_metrics(metric: &ResourceMetrics) -> ResourceMetrics {
-        ResourceMetrics {
+    fn clone_metrics(mut metric: ResourceMetrics<'_>) -> ResourceMetricsData {
+        let mut scope_metrics = Vec::new();
+        while let Some(mut scope_metric) = metric.scope_metrics.next() {
+            let mut metrics = Vec::new();
+            while let Some(metric) = scope_metric.metrics.next() {
+                metrics.push(Metric {
+                    name: metric.name.clone(),
+                    description: metric.description.clone(),
+                    unit: metric.unit.clone(),
+                    data: Self::clone_data(&metric.data),
+                });
+            }
+            scope_metrics.push(ScopeMetricsData {
+                scope: scope_metric.scope.clone(),
+                metrics,
+            });
+        }
+        ResourceMetricsData {
             resource: metric.resource.clone(),
-            scope_metrics: metric
-                .scope_metrics
-                .iter()
-                .map(|scope_metric| ScopeMetrics {
-                    scope: scope_metric.scope.clone(),
-                    metrics: scope_metric
-                        .metrics
-                        .iter()
-                        .map(|metric| Metric {
-                            name: metric.name.clone(),
-                            description: metric.description.clone(),
-                            unit: metric.unit.clone(),
-                            data: Self::clone_data(&metric.data),
-                        })
-                        .collect(),
-                })
-                .collect(),
+            scope_metrics,
         }
     }
 
@@ -237,7 +261,7 @@ impl InMemoryMetricExporter {
 }
 
 impl PushMetricExporter for InMemoryMetricExporter {
-    async fn export(&self, metrics: &mut ResourceMetrics) -> OTelSdkResult {
+    async fn export(&self, metrics: ResourceMetrics<'_>) -> OTelSdkResult {
         self.metrics
             .lock()
             .map(|mut metrics_guard| {