+++ opentelemetry-sdk/src/metrics/mod.rs
@@ -112,11 +112,12 @@ pub enum Temporality {
 
 #[cfg(all(test, feature = "testing"))]
 mod tests {
-    use self::data::{HistogramDataPoint, ScopeMetrics, SumDataPoint};
+    use self::data::{HistogramDataPoint, SumDataPoint};
     use super::data::MetricData;
     use super::internal::Number;
+    use super::reader::ResourceMetricsData;
+    use super::reader::ScopeMetricsData;
     use super::*;
-    use crate::metrics::data::ResourceMetrics;
     use crate::metrics::internal::AggregatedMetricsAccess;
     use crate::metrics::InMemoryMetricExporter;
     use crate::metrics::InMemoryMetricExporterBuilder;
@@ -2892,9 +2893,9 @@ mod tests {
     }
 
     fn find_scope_metric<'a>(
-        metrics: &'a [ScopeMetrics],
+        metrics: &'a [ScopeMetricsData],
         name: &'a str,
-    ) -> Option<&'a ScopeMetrics> {
+    ) -> Option<&'a ScopeMetricsData> {
         metrics
             .iter()
             .find(|&scope_metric| scope_metric.scope.name() == name)
@@ -2905,7 +2906,7 @@ mod tests {
         meter_provider: SdkMeterProvider,
 
         // Saving this on the test context for lifetime simplicity
-        resource_metrics: Vec<ResourceMetrics>,
+        resource_metrics: Vec<ResourceMetricsData>,
     }
 
     impl TestContext {